Added specific class for MouseManager
