+++ src/components/widgets/ConfirmPopup.ts
@@ -24,12 +24,9 @@ export class ConfirmPopup extends ButtonPopup {
         super(id, title, message, ["Yes", "No"])
 
         super.keyListner = (_str: string, key : KeyListenerArgs) => {
-            if (key.code && key.code === "[M") {
-                this.remainingMouseFrames = 3
-                return
-            }
-            if (this.remainingMouseFrames > 0) {
-                this.remainingMouseFrames--
+            const checkResult = this.CM.mouse.isMouseFrame(key.code, this.remainingMouseFrames)
+            if (typeof checkResult === "number") {
+                this.remainingMouseFrames = checkResult
                 return
             }
             switch (key.name) {
