+++ src/components/widgets/OptionPopup.ts
@@ -64,12 +64,9 @@ export class OptionPopup extends EventEmitter {
      * @memberof OptionPopup
      */
     keyListner(_str: string, key: KeyListenerArgs) {
-        if (key.code && key.code === "[M") {
-            this.remainingMouseFrames = 3
-            return
-        }
-        if (this.remainingMouseFrames > 0) {
-            this.remainingMouseFrames--
+        const checkResult = this.CM.mouse.isMouseFrame(key.code, this.remainingMouseFrames)
+        if (typeof checkResult === "number") {
+            this.remainingMouseFrames = checkResult
             return
         }
         switch (key.name) {
